+++ CHANGELOG.md
@@ -22,8 +22,14 @@ TrackGraph.alter() # Comment regarding the change
 - Decoding
 
     - Fix edge case errors in spike time loading #1083
+
 - Linearization
+
     - Add edge_map parameter to LinearizedPositionV1 #1091
+
+- Position
+
+    - Fix video directory bug in `DLCPoseEstimationSelection` #1103
 
 - Spike Sorting